Update README and modify greeting in main.c
The main.c file has been updated to accommodate alternate naming conventions in greetings. This change now allows the program to take an argument input from the command line and use that as the greeting name. The README file was also updated to document these changes and how to use the modified feature.